Lagere highlightText und escapeHtml aus
