+++ web/hla_translation_tool.html
@@ -664,6 +664,7 @@
 
     <script src="https://www.youtube.com/iframe_api"></script>
     <script src="src/colorUtils.js"></script>
+    <script src="src/textUtils.js"></script>
     <script src="src/main.js"></script>
     <script src="renderer.js"></script>
 </body>
